fix filestructure and requires
