+++ lib/csp/directive_value.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module CSP::DirectiveValue
+  require_relative "directive_value/base"
+  require_relative "directive_value/default"
+  require_relative "directive_value/sandbox"
+  require_relative "directive_value/source"
+  require_relative "directive_value/serialized_source_list"
+  require_relative "directive_value/token"
+
+  ParseError = Class.new(StandardError)
+  InvalidSource = Class.new(StandardError)
+end
