+++ lib/csp/directive_value/default.rb
@@ -1,9 +1,5 @@
 # frozen_string_literal: true
 
-require_relative "./directive_value"
-require_relative "./base"
-require_relative "../grammar"
-
 class CSP::DirectiveValue::Default < CSP::DirectiveValue::Base
   private
