+++ lib/csp/directive_value/source.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module CSP::DirectiveValue::Source
+  require_relative "source/base"
+  require_relative "source/hash"
+  require_relative "source/host"
+  require_relative "source/http_host"
+  require_relative "source/keyword"
+  require_relative "source/nonce"
+  require_relative "source/none"
+  require_relative "source/scheme"
+end
