+++ lib/csp/directive_value/source/hash.rb
@@ -1,9 +1,5 @@
 # frozen_string_literal: true
 
-require_relative "./source"
-require_relative "./base"
-require_relative "../../grammar"
-
 class CSP::DirectiveValue::Source::Hash < CSP::DirectiveValue::Source::Base
   def algorithm
     @match[:algorithm]
