+++ lib/csp/directive_value/source/http_host.rb
@@ -1,9 +1,5 @@
 # frozen_string_literal: true
 
-require_relative "./source"
-require_relative "../../grammar"
-require_relative "./host"
-
 class CSP::DirectiveValue::Source::HttpHost < CSP::DirectiveValue::Source::Host
   private
