+++ lib/csp/directive_value/source/none.rb
@@ -1,9 +1,5 @@
 # frozen_string_literal: true
 
-require_relative "./source"
-require_relative "./base"
-require_relative "../../grammar"
-
 class CSP::DirectiveValue::Source::None < CSP::DirectiveValue::Source::Base
   private
