+++ lib/csp_parser.rb
@@ -1,19 +1,22 @@
 # frozen_string_literal: true
 
-require_relative "./csp"
-require_relative "./serialized_policy"
-require_relative "./directive_value/source/http_host"
+module CSP
+  require_relative "csp/directive_value"
+  require_relative "csp/directive"
+  require_relative "csp/grammar"
+  require_relative "csp/serialized_policy"
 
-class CSP::Parser
-  class << self
-    def parse(csp_policy_str)
-      CSP::SerializedPolicy.new(csp_policy_str)
-    end
+  class Parser
+    class << self
+      def parse(csp_policy_str)
+        CSP::SerializedPolicy.new(csp_policy_str)
+      end
 
-    def valid_http_host_source?(http_host_str)
-      !!CSP::DirectiveValue::Source::HttpHost.new(http_host_str)
-    rescue
-      false
+      def valid_http_host_source?(http_host_str)
+        !!CSP::DirectiveValue::Source::HttpHost.new(http_host_str)
+      rescue
+        false
+      end
     end
   end
 end
