--- lib/directive_value/source/source.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-require_relative "../directive_value"
-
-module CSP::DirectiveValue::Source
-end
