+++ spec/directive_value/source/none_spec.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require_relative "../../../lib/directive_value/source/none"
-
 describe CSP::DirectiveValue::Source::None do
   it "parses right none" do
     nonce = CSP::DirectiveValue::Source::None.new("'none'")
