+++ spec/spec_helper.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require "simplecov"
+require "csp_parser"
 
 SimpleCov.start do
   enable_coverage :branch
